Fixed bug where mapping results sometimes wouldn't display. This would occur due to a chromosome (in this case the last) not having any markers.
Improved the way plink gets its path/command to use a method similar to the
one Pjotr used with pylmm. I'll also do this for the other mapping methods.

Fixed issue where the Y axis would always say LOD score. It now says LRS for
mapping methods that return LRS

Switched interval mapping (qtl reaper) to use the marker_regression template and removed the interval_mapping template (since it's unnecessary)

Some commented out changes remain (in show_trait_mapping_tools and create_lodchart) from when I was attempting to open the mapping results
in a new page. I had resolved every issue but the mapping javascript (lod_chart) not
being able to access js_data (which has all the result data; markers, p-values, etc). I'm pretty
sure that this is because js_data was inserted into the html after the page was loaded while
the chart code ran immediately. I experimented with adding a short timeout to the mapping
javascript and data table javascript, but while it worked for the table it did not work for the
mapping figure. I don't know why this is.