Mapping methods now check for existing genotype files.
Currently we still usually get our samplelists from the genofile. This is
dumb because it results in us having a bunch of "dummy" genofiles for certain
data sets (seems to be mostly human ones). This means that checking for the
genofile alone isn't enough to determine if a mapping method should exist
for a given group

I wrote some code that will instead get the samplelist from the plink .fam file
for some of these groups/datasets (if the .fam file exists). Ideally I would like to remove all of the dummy
.geno files, but we can't yet do so because it's currently the only place we seem to be storing
the sample list for some groups.

I also moved gemma into the plink directory to get it out of the git tree.
Since it uses the same files as plink, it doesn't make sense for it
to be in its own separate directory